requirements.txt
requests
python-dotenv
psycopg2-binary
google-api-python-client
